Added the ability to change the navigation bar when changing the user role. Changed controllers.
