+++ Q1TimeS/Q1TimeS.csproj
@@ -25,4 +25,8 @@
     <PackageReference Include="Pomelo.EntityFrameworkCore.MySql" Version="7.0.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Views\General\" />
+  </ItemGroup>
+
 </Project>
