--- Q1TimeS/Views/General/SurveyInfoPage.cshtml
@@ -1,4 +1,0 @@
-﻿@{
-    ViewData["Title"] = "Опросы";
-    ViewData["StyleName"] = "surveys.css";
-}
