+++ Q1TimeS/Views/Home/Index.cshtml
@@ -9,7 +9,7 @@
     <h1 class="choose-title">Потребитель или создатель?</h1>
     <div class="choose-action-block">
         @if (activeCard != "admin"){
-            <div class="action-block" id="general-block" name="general-block">
+            <div class="action-block" id="user-block" name="user-block">
                 <i class='bx bx-server'></i>
                 <div class="block-description"><p>Подключитесь к <br>опросу и помогите вашим коллегам принять верное решение</p></div>
             </div>
