+++ Q1TimeS/Views/Shared/_AdminPartial.cshtml
@@ -1,0 +1,6 @@
+﻿<li class="nav-item">
+    <a class="nav-link mx-lg-2 text-light" asp-area="" asp-controller="Admin" asp-action="Statistics">Статистика</a>
+</li>
+<li class="nav-item">
+    <a class="nav-link mx-lg-2 text-light" asp-area="" asp-controller="Admin" asp-action="Workshop">Мастерская опросов</a>
+</li>
