+++ Q1TimeS/Views/Shared/_UserPartial.cshtml
@@ -1,0 +1,3 @@
+﻿<li class="nav-item">
+    <a class="nav-link mx-lg-2 text-light" asp-area="" asp-controller="User" asp-action="SurveysList">Список опросов</a>
+</li>
